Task 0007 complete.
This was a rather long one. I wish I had broken it down and made more commits, but here we are. All test passing - the end-to-end tests were more complex this time which is why it took so long. Scheduled Event list display format is a bit ridiculous, but it's a starting point.